# senai_crm/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SenAI CRM",
    "description": "AI-powered email operations and triage system",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/ingest": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest a single customer email",
        "description": "Ingest one email, run the rule engine, classify with LLM if eligible,\nand persist results. Idempotent: duplicate message_id returns 200 with\nalready_exists=True and the original persisted data.",
        "operationId": "ingest_email_api_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "List ingested emails",
        "operationId": "list_emails_api_emails_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/emails/{email_id}": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Email detail with actions",
        "operationId": "get_email_api_emails__email_id__get",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/status/{job_id}": {
      "get": {
        "tags": [
          "ingest"
        ],
        "summary": "Processing status by message_id",
        "operationId": "get_job_status_api_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "proposed_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Content"
          },
          "is_approved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Approved"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "agent_reasoning_log": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Reasoning Log"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action_type",
          "proposed_content",
          "is_approved",
          "approved_by",
          "agent_reasoning_log",
          "created_at"
        ],
        "title": "ActionSummary"
      },
...
      "EmailDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "requires_human": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires Human"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "raw_entities": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Entities"
          },
          "rule_flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Flags"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSummary"
            },
            "type": "array",
            "title": "Actions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "sender",
          "subject",
          "body",
          "timestamp",
          "status",
          "category",
          "urgency",
          "sentiment_score",
          "priority_score",
          "requires_human",
          "confidence",
          "raw_entities",
          "rule_flags",
          "thread_id",
          "actions"
        ],
        "title": "EmailDetail"
      },
...
      "EmailIngestRequest": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "sender",
          "timestamp",
          "thread_id"
        ],
        "title": "EmailIngestRequest"
      },
      "EmailIngestResponse": {
        "properties": {
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "already_exists": {
            "type": "boolean",
            "title": "Already Exists"
          },
          "email_id": {
            "type": "integer",
            "title": "Email Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "priority_score": {
            "type": "integer",
            "title": "Priority Score"
          },
          "rule_flags": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rule Flags"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "rag_chunks_used": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rag Chunks Used",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "message_id",
          "already_exists",
          "email_id",
          "thread_id",
          "priority_score",
          "rule_flags",
          "status"
        ],
        "title": "EmailIngestResponse"
      },
      "EmailListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "sentiment_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sentiment Score"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "requires_human": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires Human"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message_id",
          "sender",
          "subject",
          "timestamp",
          "status",
          "category",
          "urgency",
          "sentiment_score",
          "priority_score",
          "requires_human",
          "thread_id"
        ],
        "title": "EmailListItem"
      },
      "EmailListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "emails": {
            "items": {
              "$ref": "#/components/schemas/EmailListItem"
            },
            "type": "array",
            "title": "Emails"
          }
        },
        "type": "object",
        "required": [
          "total",
          "emails"
        ],
        "title": "EmailListResponse"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error_code": {
            "type": "string",
            "title": "Error Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "description": "Consistent error response shape for all API endpoints.\nEvaluators check that malformed payloads return this structure."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "email_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "urgency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Urgency"
          },
          "requires_human": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires Human"
          },
          "found": {
            "type": "boolean",
            "title": "Found"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "email_id",
          "status",
          "category",
          "urgency",
          "requires_human",
          "found"
        ],
        "title": "JobStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}